Authentication/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
pytest-django==4.5.2
pytest==7.3.1
psycopg2-binary
virtualenv